fix: cast row_number bigint to int for bench_typo_name
row_number() over () returns bigint, but bench_typo_name takes an int
position argument, so the call failed to resolve. Cast at the call
site.

+++ bench/perturbed.sql
@@ -229,7 +229,7 @@ named as (
   select
     base_id, kind, postal_code, phone, latitude, longitude, rn, orig_street_no, street_name,
     case
-      when kind in ('typo', 'typo_no_phone') then bench_typo_name(name_cased, rn)
+      when kind in ('typo', 'typo_no_phone') then bench_typo_name(name_cased, rn::int)
       when kind in ('abbrev', 'abbrev_no_phone') then
         regexp_replace(
           regexp_replace(
